requirements.txt
streamlit
google-generativeai
pandas
plotly
pypdf
python-dotenv
